+++ app/src/main/java/com/example/android_technique_collection/data/remote/UnsplashApi.kt
@@ -1,7 +1,7 @@
 package com.example.android_technique_collection.data.remote
 
 import com.example.android_technique_collection.UNSPLASH_API_KEY
-import com.example.android_technique_collection.domain.model.unsplash.SearchPhotoResult
+import com.example.android_technique_collection.domain.model.unsplash.SearchPhotosResult
 import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.Query
@@ -10,5 +10,5 @@ interface UnsplashApi {
 
     @Headers("Authorization: Client-ID $UNSPLASH_API_KEY")
     @GET("search/photos")
-    suspend fun searchPhotos(@Query("query") query: String): SearchPhotoResult
+    suspend fun searchPhotos(@Query("query") query: String): SearchPhotosResult
 }
