+++ app/src/main/java/com/example/android_technique_collection/data/repository/PhotoRepositoryImpl.kt
@@ -1,14 +1,14 @@
 package com.example.android_technique_collection.data.repository
 
 import com.example.android_technique_collection.data.remote.UnsplashApi
-import com.example.android_technique_collection.domain.model.unsplash.SearchPhotoResult
+import com.example.android_technique_collection.domain.model.unsplash.SearchPhotosResult
 import com.example.android_technique_collection.domain.repository.PhotoRepository
 import javax.inject.Inject
 
 class PhotoRepositoryImpl @Inject constructor(
     private val unsplashApi: UnsplashApi
 ): PhotoRepository {
-    override suspend fun searchPhotos(query: String): SearchPhotoResult {
+    override suspend fun searchPhotos(query: String): SearchPhotosResult {
         return unsplashApi.searchPhotos(query)
     }
 }
