--- app/src/main/java/com/example/android_technique_collection/domain/model/unsplash/Links.kt
@@ -1,12 +1,0 @@
-package com.example.android_technique_collection.domain.model.unsplash
-
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Links(
-    val download: String?,
-    val html: String?,
-    val self: String?
-)
