--- app/src/main/java/com/example/android_technique_collection/domain/model/unsplash/Result.kt
@@ -1,26 +1,0 @@
-package com.example.android_technique_collection.domain.model.unsplash
-
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Result(
-    @Json(name = "blur_hash")
-    val blurHash: String?,
-    val color: String?,
-    @Json(name = "created_at")
-    val createdAt: String?,
-    @Json(name = "current_user_collections")
-    val currentUserCollections: List<Any>?,
-    val description: String?,
-    val height: Int?,
-    val id: String?,
-    @Json(name = "liked_by_user")
-    val likedByUser: Boolean?,
-    val likes: Int?,
-    val links: Links?,
-    val urls: Urls?,
-    val user: User?,
-    val width: Int?
-)
