+++ app/src/main/java/com/example/android_technique_collection/domain/model/unsplash/SearchPhotoResult.kt
@@ -6,8 +6,37 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class SearchPhotoResult(
-    val results: List<Result>?,
-    val total: Int?,
-    @Json(name = "total_pages")
-    val totalPages: Int?
+    @Json(name = "blur_hash")
+    val blurHash: String?,
+    val color: String?,
+    @Json(name = "created_at")
+    val createdAt: String?,
+    @Json(name = "current_user_collections")
+    val currentUserCollections: List<Any>?,
+    val description: String?,
+    val height: Int?,
+    val id: String?,
+    @Json(name = "liked_by_user")
+    val likedByUser: Boolean?,
+    val likes: Int?,
+    val links: Links?,
+    val urls: Urls?,
+    val user: User?,
+    val width: Int?
+)
+
+@JsonClass(generateAdapter = true)
+data class Links(
+    val download: String?,
+    val html: String?,
+    val self: String?
+)
+
+@JsonClass(generateAdapter = true)
+data class Urls(
+    val full: String?,
+    val raw: String?,
+    val regular: String?,
+    val small: String?,
+    val thumb: String?
 )
