--- app/src/main/java/com/example/android_technique_collection/domain/model/unsplash/Urls.kt
@@ -1,14 +1,0 @@
-package com.example.android_technique_collection.domain.model.unsplash
-
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Urls(
-    val full: String?,
-    val raw: String?,
-    val regular: String?,
-    val small: String?,
-    val thumb: String?
-)
