+++ app/src/main/java/com/example/android_technique_collection/domain/repository/PhotoRepository.kt
@@ -1,8 +1,8 @@
 package com.example.android_technique_collection.domain.repository
 
-import com.example.android_technique_collection.domain.model.unsplash.SearchPhotoResult
+import com.example.android_technique_collection.domain.model.unsplash.SearchPhotosResult
 
 interface PhotoRepository {
 
-    suspend fun searchPhotos(query: String): SearchPhotoResult
+    suspend fun searchPhotos(query: String): SearchPhotosResult
 }
